feat(roles): police and doctor are single-instance; drop night majority
경찰·의사를 다른 역할처럼 1인 역할로 전환:
- 설정이 수 대신 활성화 토글로 동작 (웹 설정 Bool 필드, /마피아설정
  bool 파라미터, 설정 요약 표기 갱신). 기존 config의 다인 값은 로드 시
  0/1로 눌러 읽는다.
- 경찰 조사 과반 집계 제거: 제출된 조사가 그대로 결과가 되고, "조사
  투표 대상" → "조사 대상", "과반에 도달하지 못해..." 문구 삭제.
  경찰·대상이 그 밤에 죽어도 이미 성립한 결과는 유지된다.
- 마피아 공격 과반(다인 정상 역할)은 그대로다.

+++ src/commands.rs
@@ -1718,9 +1718,8 @@ pub async fn handle_night_action(
                 return Ok(());
             }
         };
-        // 경찰은 대상을 고른 즉시 조사 결과를 본다. 대상을 바꾸면 바꾼 대상의 결과가
-        // 다시 붙는다. 밤이 끝날 때 나오는 결과는 경찰이 여러 명일 때의 과반 결과이므로
-        // 그대로 남겨둔다.
+        // 경찰은 대상을 고른 즉시 조사 결과를 본다. 밤이 끝날 때 나오는 결과는
+        // 같은 내용의 재안내다.
         let message = match running_write.game.police_result_for_actor(actor_id) {
             Some(result) => format!("{message}\n{result}"),
             None => message,
@@ -3328,8 +3327,8 @@ pub async fn reset_leaderboard(ctx: Context<'_>) -> Result<(), Error> {
 pub async fn configure_game(
     ctx: Context<'_>,
     #[description = "마피아 수"] mafia: Option<u32>,
-    #[description = "의사 수"] doctor: Option<u32>,
-    #[description = "경찰 수"] police: Option<u32>,
+    #[description = "의사 활성화 여부"] doctor: Option<bool>,
+    #[description = "경찰 활성화 여부"] police: Option<bool>,
     #[description = "시민 특수룰 수"] citizen_special: Option<u32>,
     #[description = "마피아 특수룰 수"] mafia_special: Option<u32>,
     #[description = "중립 특수룰 수"] neutral_special: Option<u32>,
@@ -3373,10 +3372,10 @@ pub async fn configure_game(
         config_write.default_mafia_count = value;
     }
     if let Some(value) = doctor {
-        config_write.default_doctor_count = value;
+        config_write.default_doctor_count = u32::from(value);
     }
     if let Some(value) = police {
-        config_write.default_police_count = value;
+        config_write.default_police_count = u32::from(value);
     }
     if let Some(value) = citizen_special {
         config_write.citizen_special_count = value;
